# services/api/openapi.yaml
openapi: 3.1.0
info:
  title: AI Bookworm API
  version: 0.1.0
  description: Mandatory paths/schemas per MASTER-BUILD-SPEC section 9. All endpoints versioned under /v1.
servers:
  - url: http://localhost:3001
...
paths:
...
  /v1/referrals/claim:
    post:
      tags: [v1]
      operationId: claimReferral
      summary: Attribute the caller to a referral code at signup
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [code], properties: { code: { type: string } } }
      responses:
        "201": { description: Attributed }
        "200": { description: Already attributed }
        "404": { $ref: "#/components/responses/Error" }
        "422": { description: Self-referral / inactive code }
  /v1/referrals/qualify:
    post:
      tags: [internal]
      operationId: qualifyReferral
      summary: Qualify a referral after a qualified action (service token). Anti-fraud may hold the reward for review. Idempotent.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [referredUserId], properties: { referredUserId: { type: string, format: uuid } } }
      responses:
        "200": { description: Qualified (rewarded, held, or duplicate no-op) }
        "401": { $ref: "#/components/responses/Error" }
  /v1/referrals/{id}/review:
    post:
      tags: [v1]
      operationId: reviewReferral
      summary: Admin review of a held referral (approve posts the reward once)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object, required: [action], properties: { action: { type: string, enum: [approve, reject] } } }
      responses:
        "200": { description: Reviewed }
        "403": { $ref: "#/components/responses/Error" }
  /v1/referrals/{id}/reverse:
    post:
      tags: [v1]
      operationId: reverseReferral
      summary: Admin reversal (fraud/chargeback) — compensating negative ledger entry
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200": { description: Reversed }
        "403": { $ref: "#/components/responses/Error" }
        "422": { $ref: "#/components/responses/Error" }
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code: { type: string }
            message: { type: string }
            requestId: { type: string }
            details: { type: object }
...
  responses:
    Error:
      description: Standard error envelope
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
security:
  - bearerAuth: []